+++ ai.h
@@ -1,0 +1,8 @@
+#ifndef AI_H
+#define AI_H
+
+#include "widget.h"
+#include <math.h>
+#include <QWidget>
+
+#endif // AI_H
